Move ChangeLock to shared package
