+++ collect_app/src/main/java/org/odk/collect/android/application/Collect.java
@@ -41,7 +41,7 @@ import org.odk.collect.projects.ProjectsDependencyComponent;
 import org.odk.collect.projects.ProjectsDependencyComponentProvider;
 import org.odk.collect.projects.ProjectsDependencyModule;
 import org.odk.collect.projects.ProjectsRepository;
-import org.odk.collect.shared.Md5;
+import org.odk.collect.shared.strings.Md5;
 import org.odk.collect.shared.Settings;
 import org.odk.collect.strings.LocalizedApplication;
