--- collect_app/src/main/java/org/odk/collect/android/backgroundwork/ChangeLock.java
@@ -1,8 +1,0 @@
-package org.odk.collect.android.backgroundwork;
-
-import java.util.function.Function;
-
-public interface ChangeLock {
-
-    <T> T withLock(Function<Boolean, T> function);
-}
