+++ collect_app/src/main/java/org/odk/collect/android/metadata/SharedPreferencesInstallIDProvider.java
@@ -2,7 +2,7 @@ package org.odk.collect.android.metadata;
 
 import org.odk.collect.shared.Settings;
 
-import static org.odk.collect.shared.RandomString.randomString;
+import static org.odk.collect.shared.strings.RandomString.randomString;
 
 public class SharedPreferencesInstallIDProvider implements InstallIDProvider {
