+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/FormMetadataPreferencesFragment.java
@@ -16,7 +16,7 @@ import org.odk.collect.android.listeners.PermissionListener;
 import org.odk.collect.android.logic.PropertyManager;
 import org.odk.collect.android.permissions.PermissionsProvider;
 import org.odk.collect.android.utilities.ToastUtils;
-import org.odk.collect.shared.Validator;
+import org.odk.collect.shared.strings.Validator;
 
 import javax.inject.Inject;
