+++ collect_app/src/main/java/org/odk/collect/android/utilities/EncryptionUtils.java
@@ -32,7 +32,7 @@ import org.odk.collect.android.provider.FormsProviderAPI;
 import org.odk.collect.android.provider.InstanceProviderAPI;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.instances.Instance;
-import org.odk.collect.shared.Md5;
+import org.odk.collect.shared.strings.Md5;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
