+++ collect_app/src/main/java/org/odk/collect/android/utilities/FormDefCache.java
@@ -4,7 +4,7 @@ import org.javarosa.core.model.FormDef;
 import org.javarosa.core.util.externalizable.ExtUtil;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
-import org.odk.collect.shared.Md5;
+import org.odk.collect.shared.strings.Md5;
 
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
