+++ collect_app/src/main/java/org/odk/collect/android/utilities/FormsDirDiskFormsSynchronizer.java
@@ -10,8 +10,8 @@ import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.storage.StorageSubdirectory;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.FormsRepository;
-import org.odk.collect.shared.Md5;
-import org.odk.collect.shared.Validator;
+import org.odk.collect.shared.strings.Md5;
+import org.odk.collect.shared.strings.Validator;
 
 import java.io.File;
 import java.util.ArrayList;
