+++ collect_app/src/test/java/org/odk/collect/android/backgroundwork/SyncFormsTaskSpecTest.java
@@ -19,6 +19,7 @@ import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.support.CollectHelpers;
 import org.odk.collect.android.utilities.FormsRepositoryProvider;
 import org.odk.collect.android.utilities.InstancesRepositoryProvider;
+import org.odk.collect.shared.locks.ChangeLock;
 
 import java.util.HashMap;
