+++ collect_app/src/test/java/org/odk/collect/android/external/ExternalDataReaderTest.java
@@ -15,7 +15,7 @@ import org.odk.collect.android.support.CollectHelpers;
 import org.odk.collect.android.utilities.CustomSQLiteQueryBuilder;
 import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.android.utilities.SQLiteUtils;
-import org.odk.collect.shared.Md5;
+import org.odk.collect.shared.strings.Md5;
 
 import java.io.BufferedWriter;
 import java.io.File;
