+++ collect_app/src/test/java/org/odk/collect/android/formmanagement/FormUpdateCheckerTest.kt
@@ -23,13 +23,13 @@ import org.odk.collect.android.notifications.Notifier
 import org.odk.collect.android.preferences.keys.GeneralKeys
 import org.odk.collect.android.provider.FormsProviderAPI.CONTENT_URI
 import org.odk.collect.android.storage.StorageSubdirectory
-import org.odk.collect.android.support.BooleanChangeLock
 import org.odk.collect.forms.FormListItem
 import org.odk.collect.forms.FormSource
 import org.odk.collect.forms.FormSourceException
 import org.odk.collect.formstest.FormUtils
 import org.odk.collect.projects.Project
-import org.odk.collect.shared.Md5.getMd5Hash
+import org.odk.collect.shared.strings.Md5.getMd5Hash
+import org.odk.collect.testshared.BooleanChangeLock
 
 @RunWith(AndroidJUnit4::class)
 class FormUpdateCheckerTest {
