+++ collect_app/src/test/java/org/odk/collect/android/openrosa/OpenRosaGetRequestTest.java
@@ -5,7 +5,7 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.odk.collect.android.openrosa.support.MockWebServerRule;
-import org.odk.collect.shared.Md5;
+import org.odk.collect.shared.strings.Md5;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
