+++ collect_app/src/test/java/org/odk/collect/android/provider/FormsProviderTest.java
@@ -18,7 +18,7 @@ import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.support.CollectHelpers;
 import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.formstest.FormUtils;
-import org.odk.collect.shared.Md5;
+import org.odk.collect.shared.strings.Md5;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
