+++ projects/src/test/java/org/odk/collect/projects/InMemProjectsRepositoryTest.kt
@@ -1,6 +1,6 @@
 package org.odk.collect.projects
 
-import org.odk.collect.shared.UUIDGenerator
+import org.odk.collect.shared.strings.UUIDGenerator
 
 class InMemProjectsRepositoryTest : ProjectsRepositoryTest() {
     override fun buildSubject(): ProjectsRepository {
