+++ projects/src/test/java/org/odk/collect/projects/SharedPreferencesProjectsRepositoryTest.kt
@@ -3,7 +3,7 @@ package org.odk.collect.projects
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.gson.Gson
 import org.junit.runner.RunWith
-import org.odk.collect.shared.UUIDGenerator
+import org.odk.collect.shared.strings.UUIDGenerator
 import org.odk.collect.testshared.InMemSettings
 
 @RunWith(AndroidJUnit4::class)
