+++ shared/src/main/java/org/odk/collect/shared/TempFiles.kt
@@ -1,5 +1,6 @@
 package org.odk.collect.shared
 
+import org.odk.collect.shared.strings.RandomString
 import java.io.File
 
 object TempFiles {
