+++ shared/src/main/java/org/odk/collect/shared/strings/Md5.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.shared
+package org.odk.collect.shared.strings
 
 import java.io.File
 import java.io.FileInputStream
