+++ shared/src/main/java/org/odk/collect/shared/strings/UUIDGenerator.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.shared
+package org.odk.collect.shared.strings
 
 import java.util.UUID
