+++ shared/src/main/java/org/odk/collect/shared/strings/Validator.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.odk.collect.shared
+package org.odk.collect.shared.strings
 
 import java.util.regex.Pattern
