+++ shared/src/test/java/org/odk/collect/shared/Md5Test.kt
@@ -2,6 +2,7 @@ package org.odk.collect.shared
 
 import org.junit.Assert
 import org.junit.Test
+import org.odk.collect.shared.strings.Md5
 import java.io.File
 import java.io.FileWriter
